sage.rings.function_field: Move some imports into methods
